Before:
```
BenchmarkLogsClone-16    	    2022	    526826 ns/op
BenchmarkMetricsClone-16    	      92	  11530154 ns/op
BenchmarkTracesClone-16    	     237	   4822875 ns/op
```

After:
```
BenchmarkLogsClone-16    	   17643	     70484 ns/op
BenchmarkMetricsClone-16    	     516	   2226758 ns/op
BenchmarkTracesClone-16    	    1286	    853794 ns/op
```

Signed-off-by: Bogdan Drutu <bogdandrutu@gmail.com>