@bogdandrutu 2020-10-22
The tests that I added for clone fill all the fields, and verify all of them are "copied".

In terms of reference, we do use same references to the immutable objects.